fix: Shorten scope of loop to get numbers before printing
Instead of printing the numbers inside of the while loop that counts from 0 to n-1, it will print outside of the loop. The first while loop will just get the numbers entered by the user and put them in the list. After that, it will print all the numbers inside the list.