fix: Fix LLISTA_insertDavant function
Due to the changes made in creating the list, the way we insert a node after PDI also changed. Pointer sig of auxiliar node will be pointing to where PDI sig was pointing and pointer ant of auxiliar node will be pointing to PDI sig points at.
For inserting a node for the first time, this will insert a node in between ghost nodes PRI and ULT.