fix: Change implementation of LLISTA_insertDavant function
The new functionality of this is that it checks if the list is empty. If it is, the PDI will be the First ghost node. Then it inserts it to the right.
If it is not empty, it will go to the first node and start from there. It will check if the node's element is smaller than the element passed as parameter. In this case, it moves backward to insert to the right in a sorted manner.