fix: Fix LLISTA_insertDavant function
Add a condition that checks if PDI is pointing at the last node so it doesn't try to read and get the element in the last node. This willl give segmentation fault. So if we are already at the last node, then we insert the element passed as parameter before the last node.