fix: Fix LLISTA_insertDavant function
It entered into an infinite loop because when we moved backwards when it detected that we are at the end of the list, it never detects that we are at the end already. For that, we say that we have found the perfect position for the element passed as parameter to the function.